refactor: move PostService and FirebaseConvertable to their own files
