+++ FirebaseTestProject/Models/Post.swift
@@ -6,7 +6,6 @@
 //
 
 import Foundation
-import FirebaseFirestore
 
 struct Post: FirebaseConvertable {
     let title: String
@@ -23,39 +22,4 @@ struct Post: FirebaseConvertable {
         self.timestamp = Date()
     }
     static let testPost = Post(title: "Title", text: "Content", author: "First Last")
-}
-
-protocol FirebaseConvertable: Codable {
-    init(from jsonDict: [String: Any])
-    var jsonDict: [String: Any] { get }
-}
-
-extension FirebaseConvertable {
-    init(from jsonDict: [String: Any]) {
-        let data = try! JSONSerialization.data(withJSONObject: jsonDict)
-        let newInstance = try! JSONDecoder().decode(Self.self, from: data)
-        self = newInstance
-    }
-    var jsonDict: [String: Any] {
-        let data = try! JSONEncoder().encode(self)
-        let jsonObject = try! JSONSerialization.jsonObject(with: data)
-        return jsonObject as! [String: Any]
-    }
-}
-
-struct PostService {
-    static var postsReference: CollectionReference {
-        let db = Firestore.firestore()
-        return db.collection("posts")
-    }
-    
-    static func getPosts() async throws -> [Post] {
-        let postsSnapshots = try await postsReference.getDocuments()
-        let posts = postsSnapshots.documents.map { Post(from: $0.data()) }
-        return posts
-    }
-    
-    static func upload(_ post: Post) async throws {
-        try await postsReference.document(post.id.uuidString).setData(post.jsonDict)
-    }
 }
